# micro-optimize UUID(::AbstractString)