--- a/stdlib/UUIDs/test/runtests.jl
+++ b/stdlib/UUIDs/test/runtests.jl
@@ -62,3 +62,14 @@ u4 = uuid4()
 Random.seed!(Random.GLOBAL_RNG, 10)
 @test u1 != uuid1()
 @test u4 != uuid4()
+
+@test_throws ArgumentError UUID("22b4a8a1ae548-4eeb-9270-60426d66a48e")
+@test_throws ArgumentError UUID("22b4a8a1-e548a4eeb-9270-60426d66a48e")
+@test_throws ArgumentError UUID("22b4a8a1-e548-4eeba9270-60426d66a48e")
+@test_throws ArgumentError UUID("22b4a8a1-e548-4eeb-9270a60426d66a48e")
+str = "22b4a8a1-e548-4eeb-9270-60426d66a48e"
+@test UUID(uppercase(str)) == UUID(str)
+
+for r in rand(UInt128, 10^3)
+    @test UUID(r) == UUID(string(UUID(r)))
+end