I was profiling Pkg and saw some traces pointing to `UUID(::String)`. The current implementations allocates a brand new vector for `lowercase` (which I would guess is almost never needed) and it also does a regex call to check it the format of the string is ok. This speeds it up by a bit.

Setup:

```jl
julia> using UUIDs

julia>  a = uuid4() |> string

julia> uuid_strs = string.(UUID.(rand(UInt128, 10^6)));

julia> uuids = Vector{UUID}(undef, 10^6);
```

Master

``` 
julia> @btime UUID($a);
  319.275 ns (3 allocations: 224 bytes)

julia> @btime $uuids .= UUID.($uuid_strs)
  373.077 ms (3000000 allocations: 213.62 MiB)
```



PR
```
julia> @btime UUID($a);
  27.489 ns (0 allocations: 0 bytes)

julia> @btime $uuids .= UUID.($uuid_strs)
  111.078 ms (0 allocations: 0 bytes)
```