@KristofferC 2020-05-28
> but since we do print a lot of these, 

I think printing them would be relevant for the `string(::UUID)` function but this is more for ingesting them.