@StefanKarpinski 2020-05-28
Yeah, we do a lot of UUID parsing too.